test(gdt): fold operator-arithmetic tests into test_hypothesis_operator_interface
SonarCloud flagged 81 duplicated lines (the entire PR's new-code duplication),
all in the standalone test_hypothesis_operator_arithmetic.py whose grid/operator
helper block mirrored test_hypothesis_operator_interface.py. Move the five
arithmetic tests (subtraction A - B / A - vector, jacobian_options/invert_options
match-and-mismatch, and the ConstantOperator "zero" jacobian via A + vector) into
the operator-interface module so they reuse its existing helpers, and drop the
duplicate file. Only a small _empty_matrix_operator jacobian-target helper is
added. No test logic changes.
